raspberry-pi/dashboard/requirements.txt
Flask==3.0.0
paho-mqtt==1.6.1
python-dateutil==2.8.2

